Implement compatibility with kernels pre 3.8
Support for listing inotify watches in a /proc/{}/fdinfo/{} file
was added in linux kernel 3.8; this utility relies on that information
to count the number of watches in use.

This commit detects when the inotify info is not provided in fdinfo and
ommits the "Watches" column and stats when printing the found info.

Author-Rebase-Consent: https://No-rebase.github.io