Implement build compatibility with GCC pre 7.1
The -Wimplicit-fallthrough=2 warning was first implemented in GCC 7.1.
This would cause compiling with earlier GCC to fail.

This warning (albeit with =3 format) was included in -Wall which is more
stringent. This incantation requires fall-through case statements to be
specified case sensitively. In any case, the codebase here does not
include any fall-through waivers, so changing to -Wall does not require
any other adaptations.

This commit removes the -Wimplicit-fallthrough argument from the
Makefile, and replaces it with -Wall which is more stringent, and
allows old GCC's to successfully compile the utility.

Author-Rebase-Consent: https://No-rebase.github.io