Add option of search path
inotify-info searches the root directory in the function `find_files_in_inode_set` by default. When the root is large, it is very slow. Sometimes we know that the program is watching some files under the $HOME directory but we do not know the specific files. Under such circumstance, it might be better if we can specify a search path to reduce the searching time.